restruct gui to allow concurrent compute
This moves the computation of the Features data off to another thread
and adds the plumbing to allow the state that the UI renders to be
updated from another thread.

The result is that we show the initial boundary and then a few moments
later, after the twonets have been computed, update the gui to show
their paths.