manually swap gfx buffers
The goal is to reduce the CPU usage during rendering.  We already
know when things have changed so we'd like to avoid explicitly
re-rendering all the lines again on each render call.

Assuming that the gfx system doesn't mess with the buffers while
we're not looking, we can simply keep the current buffer displayed
until something changes.

This does mean that the RenderLines struct isn't really needed
any more, but I'm keeping it around for now in case I do run
into problems with this approach in the future.